Rename services.js component to industriesWeServe.js component
